update to objective
It would be nice to generate non-lattice/grid triangulation to reduce geometrical bias, such that the discretization pattern does not constraint or steer the dynamics of surface properties during simulation. Also consider running simulations on surfaces of similar shape but with different discretization patterns, to demonstrate dynamics is independent of the cutting.